backend/requirements.txt
fastapi
uvicorn
ollama
python-dotenv
pydantic
langchain-community
chromadb
llama-index-core
llama-index-embeddings-ollama
llama-index-llms-ollama
pypdf
python-multipart
langchain-cli
rank_bm25
unstructured
langchain-ollama
rank_bm25